requirements.txt
alt-profanity-check==1.4.1.post1
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
constantly==23.10.4
contextlib2==21.6.0
cryptography==42.0.3
daphne==4.1.0
Django==5.0.2
django-ninja==1.1.0
django-ninja-extra==0.20.4
django-ninja-jwt==5.3.0
dnspython==2.6.1
email_validator==2.1.1
hyperlink==21.0.0
idna==3.6
importlib-metadata==4.8.3
incremental==22.10.0
iniconfig==2.0.0
injector==0.21.0
joblib==1.4.0
Markdown==3.3.7
martor==1.6.28
ninja-schema==0.13.6
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pluggy==1.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pyOpenSSL==24.0.0
pytest==8.1.1
pytest-django==4.8.0
reportlab==4.1.0
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.13.0
service-identity==24.1.0
setuptools==69.1.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.4.0
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
zipp==3.17.0
zope.interface==6.2
